RemovePastKeys and CleanObsoletePersistedLastSeen
