+++ src/Nethermind/Nethermind.Trie/Pruning/TrieStore.cs
@@ -513,14 +513,7 @@ namespace Nethermind.Trie.Pruning
 
                 if (_livePruningEnabled)
                 {
-                    var persistedLastSeen = DirtyNodes.PersistedLastSeen;
-                    foreach (KeyValuePair<HashAndTinyPathAndHash, long> keyValuePair in persistedLastSeen)
-                    {
-                        if (IsNoLongerNeeded(keyValuePair.Value))
-                        {
-                            persistedLastSeen.Remove(keyValuePair.Key, out _);
-                        }
-                    }
+                    DirtyNodes.CleanObsoletePersistedLastSeen();
                 }
 
                 if (candidateSets.Count > 0)
@@ -537,72 +530,7 @@ namespace Nethermind.Trie.Pruning
 
         private void RemovePastKeys(Dictionary<HashAndTinyPath, Hash256?>? persistedHashes)
         {
-            if (persistedHashes is null) return;
-
-            var persistedLastSeen = DirtyNodes.PersistedLastSeen;
-            var pastPathHash = DirtyNodes.PastPathHash;
-
-            bool CanRemove(in ValueHash256 address, TinyTreePath path, in TreePath fullPath, in ValueHash256 keccak, Hash256? currentlyPersistingKeccak)
-            {
-                // Multiple current hash that we don't keep track for simplicity. Just ignore this case.
-                if (currentlyPersistingKeccak is null) return false;
-
-                // The persisted hash is the same as currently persisting hash. Do nothing.
-                if ((ValueHash256)currentlyPersistingKeccak == keccak) return false;
-
-                // We have it in cache and it is still needed.
-                if (DirtyNodes.TryGetValue(new TrieStoreDirtyNodesCache.Key(address, fullPath, keccak.ToCommitment()), out TrieNode node) &&
-                    !IsNoLongerNeeded(node)) return false;
-
-                // We don't have it in cache, but we know it was re-committed, so if it is still needed, don't remove
-                if (persistedLastSeen.TryGetValue(new(address, in path, in keccak), out long commitBlock) &&
-                    !IsNoLongerNeeded(commitBlock)) return false;
-
-                return true;
-            }
-
-            ActionBlock<INodeStorage.WriteBatch> actionBlock =
-                new ActionBlock<INodeStorage.WriteBatch>(static (batch) => batch.Dispose());
-
-            INodeStorage.WriteBatch writeBatch = _nodeStorage.StartWriteBatch();
-            try
-            {
-                int round = 0;
-                foreach (KeyValuePair<HashAndTinyPath, Hash256> keyValuePair in persistedHashes)
-                {
-                    HashAndTinyPath key = keyValuePair.Key;
-                    if (pastPathHash.TryGet(key, out ValueHash256 prevHash))
-                    {
-                        TreePath fullPath = key.path.ToTreePath(); // Micro op to reduce double convert
-                        if (CanRemove(key.addr, key.path, fullPath, prevHash, keyValuePair.Value))
-                        {
-                            Metrics.RemovedNodeCount++;
-                            Hash256? address = key.addr == default ? null : key.addr.ToCommitment();
-                            writeBatch.Set(address, fullPath, prevHash, default, WriteFlags.DisableWAL);
-                            round++;
-                        }
-                    }
-
-                    // Batches of 256
-                    if (round > 256)
-                    {
-                        actionBlock.Post(writeBatch);
-                        writeBatch = _nodeStorage.StartWriteBatch();
-                        round = 0;
-                    }
-                }
-            }
-            catch (Exception ex)
-            {
-                if (_logger.IsError) _logger.Error($"Failed to remove past keys. {ex}");
-            }
-            finally
-            {
-                writeBatch.Dispose();
-                actionBlock.Complete();
-                actionBlock.Completion.Wait();
-                _nodeStorage.Compact();
-            }
+            DirtyNodes.RemovePastKeys(persistedHashes, _nodeStorage);
         }
 
         /// <summary>
@@ -876,7 +804,7 @@ namespace Nethermind.Trie.Pruning
             return IsNoLongerNeeded(node.LastSeen);
         }
 
-        private bool IsNoLongerNeeded(long lastSeen)
+        public bool IsNoLongerNeeded(long lastSeen)
         {
             Debug.Assert(lastSeen >= 0, $"Any node that is cache should have {nameof(TrieNode.LastSeen)} set.");
             return lastSeen < LastPersistedBlockNumber
